Standardize the usage of pitch utils
