+++ src/tools/live-set/helpers/update-live-set-helpers.js
@@ -1,4 +1,4 @@
-import { VALID_PITCH_CLASS_NAMES } from "../../../notation/pitch-class-name-to-number.js";
+import { VALID_PITCH_CLASS_NAMES } from "#src/shared/pitch.js";
 import { VALID_SCALE_NAMES } from "../../constants.js";
 import { createAudioClipInSession } from "../../shared/arrangement/arrangement-tiling.js";
